trying to lift state to parent quesitons
